+++ include/value.h
@@ -6,8 +6,8 @@
 typedef double Value;
 
 typedef struct {
-  uint32_t count;
-  uint32_t capacity;
+  size_t count;
+  size_t capacity;
   Value* values;
 } ValueArray;
